requirements.txt
Django==5.0.6
djangorestframework==3.15.1
flake8==7.0.0
psycopg2>=2.9.6,<2.10
psycopg2-binary>=2.9.6,<2.10
drf-spectacular==0.27.2
python-dotenv==1.0.1
requests==2.31.0
requests-html 
beautifulsoup4==4.12.3
lxml-html-clean==0.1.1